add displayExtensions
